requirements.txt
###
# This file is only here to make sure that something like
#
#    pip install -e .
#
# works as expected. Requirements can be found in setup.py.
###

.

setuptools~=66.1.1
OctoPrint~=1.10.2
pyserial~=3.5